added subscribed snapshot and use get request for snapshot to only receive the important data that is displayed on the homepage, instead of the whole snapshot

+++ app/services/snapshot-token.ts
@@ -8,6 +8,7 @@ import { getCircularReplacer } from 'explorviz-frontend/utils/circularReplacer';
 import { StructureLandscapeData } from 'explorviz-frontend/utils/landscape-schemes/structure-data';
 import { DynamicLandscapeData } from 'explorviz-frontend/utils/landscape-schemes/dynamic/dynamic-data';
 import { SerializedRoom } from 'collaboration/utils/web-socket-messages/types/serialized-room';
+import { Timestamp } from 'explorviz-frontend/utils/landscape-schemes/timestamp';
 
 export type SnapshotToken = {
   owner: string;
@@ -19,11 +20,26 @@ export type SnapshotToken = {
     dynamicLandscapeData: DynamicLandscapeData;
   };
   serializedRoom: SerializedRoom;
+  timestamps: { timestamps: Timestamp[] };
   camera: any;
   annotations: any;
   isShared: boolean;
+  subscribedUsers: { subscriberList: string[] };
   deleteAt: number;
   julius: any;
+};
+
+export type TinySnapshot = {
+  owner: string;
+  createdAt: number;
+  name: string;
+  landscapeToken: LandscapeToken;
+};
+
+export type SnapshotInfo = {
+  personalSnapshots: TinySnapshot[];
+  sharedSnapshots: TinySnapshot[];
+  subsricedSnapshots: TinySnapshot[];
 };
 
 const { userServiceApi } = ENV.backendAddresses;
@@ -42,15 +58,22 @@ export default class SnapshotTokenService extends Service {
   @tracked
   snapshotToken: SnapshotToken | null = null;
 
+  @tracked
+  snapshotSelected: boolean = false;
+
   // Used in landscape selection to go back to last selected snapshot
   @tracked
   latestSnapshotToken: SnapshotToken | null = null;
 
   retrieveTokens() {
-    return new Promise<SnapshotToken[]>((resolve) => {
+    return new Promise<SnapshotInfo>((resolve) => {
       const userId = encodeURI(this.auth.user?.sub || '');
       if (!userId) {
-        resolve([]);
+        resolve({
+          personalSnapshots: [],
+          sharedSnapshots: [],
+          subsricedSnapshots: [],
+        });
       }
 
       fetch(`${userServiceApi}/snapshot?owner=${userId}`, {
@@ -60,17 +83,25 @@ export default class SnapshotTokenService extends Service {
       })
         .then(async (response: Response) => {
           if (response.ok) {
-            const tokens = (await response.json()) as SnapshotToken[];
-            resolve(tokens);
+            const snapshotInfo = (await response.json()) as SnapshotInfo;
+            resolve(snapshotInfo);
           } else {
-            resolve([]);
+            resolve({
+              personalSnapshots: [],
+              sharedSnapshots: [],
+              subsricedSnapshots: [],
+            });
             this.toastHandler.showErrorToastMessage(
               'Snapshots could not be loaded.'
             );
           }
         })
         .catch(async () => {
-          resolve([]);
+          resolve({
+            personalSnapshots: [],
+            sharedSnapshots: [],
+            subsricedSnapshots: [],
+          });
           this.toastHandler.showErrorToastMessage('Server not available.');
         });
     });
@@ -141,6 +172,23 @@ export default class SnapshotTokenService extends Service {
     }
   }
 
+  async subsribe(owner: string, createdAt: number, subscriber: string) {
+    const url = `${userServiceApi}/snapshot/subscribe?owner=${owner}&createdAt=${createdAt}&subscriber=${subscriber}`;
+
+    await fetch(url, {
+      method: 'PUT',
+      headers: { 'Content-Type': 'application/json; charset=UTF-8' },
+    })
+      .then(async (response: Response) => {
+        if (!response.ok) {
+          this.toastHandler.showErrorToastMessage('Something went wrong.');
+        }
+      })
+      .catch(async () => {
+        this.toastHandler.showErrorToastMessage('Server could not be reached.');
+      });
+  }
+
   async exportFile(exportData: SnapshotToken) {
     const jsonString = `data:text/json;chatset=utf-8,${encodeURIComponent(
       JSON.stringify(exportData, getCircularReplacer())
@@ -152,17 +200,11 @@ export default class SnapshotTokenService extends Service {
     link.click();
   }
 
-  async shareSnapshot(snapshot: SnapshotToken) {
-    const snapshotToken = {
-      ...snapshot,
-      name: snapshot.name + ' (shared)',
-      isShared: true,
-    };
+  async shareSnapshot(snapshot: TinySnapshot, expDate: number) {
+    const url = `${userServiceApi}/snapshot/share?owner=${snapshot.owner}&createdAt=${snapshot.createdAt}&deleteAt=${expDate}`;
 
-    const url = `${userServiceApi}/snapshot/create`;
     await fetch(url, {
-      method: 'POST',
-      body: JSON.stringify(snapshotToken, getCircularReplacer()),
+      method: 'PUT',
       headers: { 'Content-Type': 'application/json; charset=UTF-8' },
     })
       .then(async (response: Response) => {
@@ -173,7 +215,7 @@ export default class SnapshotTokenService extends Service {
           this.toastHandler.showSuccessToastMessage(
             'Successfully shared snapshot. Snaphsot URL copied to clipboard'
           );
-        } else if (response.status === 422) {
+        } else if (response.status === 222) {
           this.toastHandler.showErrorToastMessage(
             'Snapshot could not be shared. A shared version already exists.'
           );
@@ -190,26 +232,51 @@ export default class SnapshotTokenService extends Service {
     this.router.refresh('landscapes');
   }
 
-  async deleteSnapshot(snapshot: SnapshotToken) {
-    const url = `${userServiceApi}/snapshot/delete?owner=${snapshot.owner}&createdAt=${snapshot.createdAt}&isShared=${snapshot.isShared}`;
+  async deleteSnapshot(
+    snapshot: TinySnapshot,
+    isShared: boolean,
+    subscribed: boolean
+  ) {
+    if (subscribed) {
+      const url = `${userServiceApi}/snapshot/unsubscribe?owner=${snapshot.owner}&createdAt=${snapshot.createdAt}&subscriber=${this.auth.user!.sub}`;
 
-    await fetch(url, {
-      method: 'DELETE',
-    })
-      .then(async (response: Response) => {
-        if (response.ok) {
-          this.toastHandler.showSuccessToastMessage(
-            'Successfully deleted snapshot.'
+      await fetch(url, {
+        method: 'PUT',
+        headers: { 'Content-Type': 'application/json; charset=UTF-8' },
+      }).then(async (response: Response) => {
+        if (!response.ok) {
+          this.toastHandler.showErrorToastMessage(
+            'Subscribed Snapshot could not be deleted.'
           );
         } else {
-          this.toastHandler.showErrorToastMessage(
-            'Something went wrong. Snapshot could not be deleted.'
+          this.toastHandler.showSuccessToastMessage(
+            'Subscribed Snapshot successfully deleted.'
           );
         }
-      })
-      .catch(async () => {
-        this.toastHandler.showErrorToastMessage('Server could not be reached.');
       });
+    } else {
+      const url = `${userServiceApi}/snapshot/delete?owner=${snapshot.owner}&createdAt=${snapshot.createdAt}&isShared=${isShared}`;
+
+      await fetch(url, {
+        method: 'DELETE',
+      })
+        .then(async (response: Response) => {
+          if (response.ok) {
+            this.toastHandler.showSuccessToastMessage(
+              'Successfully deleted snapshot.'
+            );
+          } else {
+            this.toastHandler.showErrorToastMessage(
+              'Something went wrong. Snapshot could not be deleted.'
+            );
+          }
+        })
+        .catch(async () => {
+          this.toastHandler.showErrorToastMessage(
+            'Server could not be reached.'
+          );
+        });
+    }
 
     this.router.refresh('landscapes');
   }
